biometric-forensic-analyzer/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

celery==5.3.6
redis==5.0.3

psycopg2-binary==2.9.9

mediapipe==0.10.14
ultralytics==8.3.110

opencv-python==4.9.0.80
Pillow==10.3.0

pyexiftool==0.5.6

numpy==1.23.5
scipy==1.10.1
PyWavelets==1.4.1

python-multipart==0.0.9